Grouped related classes as static inner classes of AttachedSobjEffect
