--- src/main/java/legend/game/submap/FootprintParticle54.java
@@ -1,6 +1,0 @@
-package legend.game.submap;
-
-/** Split off from original OrthoTrailParticle54 */
-public class FootprintParticle54 extends OrthoTrailParticle54 {
-  public int textureIndex_02;
-}
