--- src/main/java/legend/game/submap/LawPodTrailData18.java
@@ -1,22 +1,0 @@
-package legend.game.submap;
-
-import org.joml.Vector3f;
-
-public class LawPodTrailData18 {
-  public int maxCountSegments_00;
-  public int countSegments_01;
-  /** short */
-  public int fadeDelay_02;
-  /** short */
-  public int countFadeSteps_04;
-  /** short */
-  public int maxTicks_06;
-  /** short */
-  public int width_08;
-
-  public int translucency_0c;
-  /** was 3 bytes */
-  public Vector3f colour_10 = new Vector3f();
-
-  public TrailSegmentVertices14 currSegmentOriginVerts_14 = new TrailSegmentVertices14();
-}
