--- src/main/java/legend/game/submap/OrthoDustParticle54.java
@@ -1,20 +1,0 @@
-package legend.game.submap;
-
-import org.joml.Vector2f;
-
-/** Split off from original OrthoTrailParticle54 */
-public class OrthoDustParticle54 extends OrthoTrailParticle54 {
-  public float size_08;
-  public float sizeStep_0c;
-
-  public int x_18;
-  public int y_1c;
-  public final Vector2f sxy0_20 = new Vector2f();
-  public float centerX_26;
-  public final Vector2f sxy1_28 = new Vector2f();
-  public float centerY_2e;
-  public final Vector2f sxy2_30 = new Vector2f();
-  public final Vector2f sxy3_38 = new Vector2f();
-
-  public float z_4c;
-}
